♻️ Refactor: 프로덕트 이미지 컴포넌트로 분리
