+++ src/components/common/productImg/ProductImg.module.css
@@ -1,0 +1,6 @@
+.container {
+  width: 100%;
+  border-radius: 16px;
+  aspect-ratio: 1 / 1;
+  object-fit: cover;
+}
