+++ src/components/items/productCard/ProductCard.jsx
@@ -1,27 +1,15 @@
 import { Link } from 'react-router';
 import icons from '@/assets/icons/icons';
-import images from '@/assets/images/images';
+import ProductImg from '@/components/common/productImg/ProductImg';
 import { formatPrice } from '@/utils/formatPrice';
 import styles from './ProductCard.module.css';
 
 const ProductCard = ({ product }) => {
   const { id, images: productImg, name, price, favoriteCount } = product;
 
-  const imageSrc = productImg[0] || images.productEmpty;
-
-  const handleError = (e) => {
-    e.target.src = images.productEmpty;
-    e.target.onerror = null;
-  };
-
   return (
     <Link to={`/items/${id}`} className={styles.container}>
-      <img
-        className={styles.img}
-        src={imageSrc}
-        alt="상품 사진"
-        onError={handleError}
-      />
+      <ProductImg imgArray={productImg} />
       <p className={styles.title}>{name}</p>
       <span className={styles.price}>{formatPrice(price)}</span>
       <div className={styles['favorite-info-area']}>
