+++ src/components/items/productCard/ProductCard.module.css
@@ -3,13 +3,6 @@
   flex-direction: column;
   width: 100%;
 
-  .img {
-    width: 100%;
-    border-radius: 16px;
-    aspect-ratio: 1 / 1;
-    object-fit: cover;
-  }
-
   .title {
     margin-top: 16px;
     font-size: 14px;
